Change:move ol files to javascript/ol
