+++ app/javascript/ol/gsi_layer.js
