+++ app/javascript/ol/ol_control.js
