+++ app/javascript/ol/thermal_layer.js
